Moved scrollToThread from lib to inlineComment utils module
